[Backend] Refactored route to fetch the force's statuses
