+++ Backend/src/api/officers/index.ts
@@ -16,6 +16,9 @@ app.get("/", getOfficersListController);
 // Route to get the details of a specific officer
 app.get("/:nif(\\d+)", officerExistsMiddle, getOfficerDetailsController);
 
+// TODO: Add a route to check if the provided nif can be used as a new Officer, and if it belongs to a former Officer or not
+// ? Maybe this can be a HEAD route?
+
 // Route to add a new officer
 app.put("/:nif(\\d+)", addOfficerController);
