+++ Backend/src/api/util/controllers/index.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import {FORCE_HEADER} from "../../../utils/constants";
-import {forcePatents} from "../services";
-import {UtilPatentsResponse} from "@portalseguranca/api-types/util/schema";
+import {forcePatents, forceStatuses} from "../services";
+import {UtilPatentsResponse, UtilStatusesResponse} from "@portalseguranca/api-types/util/schema";
 
 export async function getPatentsController(req: express.Request, res: express.Response) {
     // Get what force the user is trying to get the patents from
@@ -12,4 +12,15 @@ export async function getPatentsController(req: express.Request, res: express.Re
 
     // Send the response to the user
     res.send({message: "Operação bem sucedida", data: result.data} as UtilPatentsResponse);
+}
+
+export async function getStatusesController(req: express.Request, res: express.Response) {
+    // Get what force the user is trying to get the patents from
+    let force = req.header(FORCE_HEADER)!;
+
+    // Call the service to get the statuses
+    const result = await forceStatuses(force);
+
+    // Send the list to the user
+    res.status(200).json({message: "Operação bem sucedida", data: result.data} as UtilStatusesResponse);
 }
